Add Overcooked observations header
- Introduced overcooked_obs.h to encapsulate observation computation functions for agents in the Overcooked environment.
- Implemented functions to compute proximity features, find the nearest empty counter, and generate a comprehensive observation vector for each agent.
- Updated overcooked.h to include the new overcooked_obs.h header, removing redundant function definitions from the main file.
